Rename commonDataclasses -> common
