+++ src/climatevision/generator/agri2018/energy_demand/co2eFromOther.py
@@ -4,7 +4,7 @@ from dataclasses import dataclass
 
 from ...inputs import Inputs
 
-from ...commonDataclasses.co2eEmissions import CO2eEmissions
+from ...common.co2eEmissions import CO2eEmissions
 
 
 @dataclass(kw_only=True)
