+++ src/climatevision/generator/agri2018/energy_demand/co2eFromSoil.py
@@ -2,7 +2,7 @@
 
 from dataclasses import dataclass
 
-from ...commonDataclasses.co2eEmissions import CO2eEmissions
+from ...common.co2eEmissions import CO2eEmissions
 
 
 @dataclass(kw_only=True)
