+++ src/climatevision/generator/agri2018/energy_source/co2eFromEnergyUse.py
@@ -2,7 +2,7 @@
 
 from dataclasses import dataclass
 
-from ...commonDataclasses.co2eEmissions import CO2eEmissions
+from ...common.co2eEmissions import CO2eEmissions
 
 
 @dataclass(kw_only=True)
