+++ src/climatevision/generator/common/energy.py
